+++ bulbs/super_features/data_serializers.py
@@ -6,9 +6,12 @@ from bulbs.utils.data_serializers import EntrySerializer
 
 
 class GuideToChildSerializer(serializers.Serializer):
-    entries = EntrySerializer(many=True, required=False)
+    entries = EntrySerializer(many=True, required=False, child_label="entry")
 
 
 class GuideToParentSerializer(serializers.Serializer):
-    sponsor_text = serializers.CharField(required=False)
-    sponsor_image = ImageFieldSerializer(required=False)
+    sponsor_brand_messaging = serializers.CharField(required=False)
+    sponsor_product_shot = ImageFieldSerializer(
+        required=False,
+        label="Sponsor Product Shot (1x1 Image)"
+    )